+++ pyfakefs/helpers.py
@@ -23,6 +23,7 @@ import os
 IS_PY2 = sys.version_info[0] < 3
 IS_PYPY = platform.python_implementation() == 'PyPy'
 IS_WIN = sys.platform == 'win32'
+IN_DOCKER = os.path.exists('/.dockerenv')
 
 try:
     text_type = unicode  # Python 2